replaced old samples with the new one
