--- newSamples/package.json
@@ -1,43 +1,0 @@
-{
-  "name": "react-vizgrammar-samples",
-  "version": "1.0.0",
-  "description": "Charting Samples for react-vizgrammar library",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "samples": "webpack-dev-server --progress"
-  },
-  "author": "",
-  "license": "Apache-2.0",
-  "dependencies": {
-    "material-ui": "^1.0.0-beta.26",
-    "rc-tree": "^1.7.10",
-    "react": "^16.2.0",
-    "react-dom": "^16.2.0",
-    "react-sortable-tree": "^1.5.4",
-    "react-vizgrammar": "^0.6.15"
-  },
-  "devDependencies": {
-    "babel-cli": "^6.26.0",
-    "babel-eslint": "^8.0.1",
-    "babel-loader": "^7.1.2",
-    "babel-preset-es2015": "^6.24.1",
-    "babel-preset-react": "^6.24.1",
-    "css-loader": "^0.28.7",
-    "eslint": "^3.19.0",
-    "eslint-config-airbnb": "^15.0.1",
-    "eslint-plugin-import": "^2.3.0",
-    "eslint-plugin-jsx-a11y": "^5.0.3",
-    "eslint-plugin-react": "^7.0.1",
-    "json-loader": "^0.5.7",
-    "log4javascript": "^1.4.15",
-    "node-sass": "^4.7.2",
-    "nodemon": "^1.12.1",
-    "react-addons-perf": "^15.4.2",
-    "react-router-dom": "^4.2.2",
-    "sass-loader": "^6.0.6",
-    "style-loader": "^0.19.0",
-    "webpack": "^3.8.1",
-    "webpack-dev-server": "^2.9.3"
-  }
-}
