+++ samples/chart-docs/AreaChartSamples.jsx
