+++ samples/chart-docs/BarChartSamples.jsx
