+++ samples/chart-docs/NumberChartSample.jsx
