+++ samples/chart-docs/ScatterPlotSample.jsx
