+++ samples/chart-docs/TableChartSamples.jsx
