+++ samples/chart-docs/helper.js
