--- samples/charts/css/default.css
@@ -1,86 +1,0 @@
-html, body, p, ol, ul, li, dl, dt, dd, blockquote, figure, fieldset, legend, textarea, pre, iframe, hr, h1, h2, h3, h4, h5, h6 {
-	margin: 0;
-	padding: 0
-}
-
-h1, h2, h3, h4, h5, h6 {
-	font-size: 100%;
-	font-weight: normal
-}
-
-ul {
-	list-style: none
-}
-
-button, input, select, textarea {
-	margin: 0
-}
-
-html {
-	box-sizing: border-box
-}
-
-*, *:before, *:after {
-	box-sizing: inherit
-}
-
-img, embed, iframe, object, audio, video {
-	height: auto;
-	max-width: 100%
-}
-
-iframe {
-	border: 0
-}
-
-table {
-	border-collapse: collapse;
-	border-spacing: 0
-}
-
-td, th {
-	padding: 0;
-	text-align: left
-}
-
-body{
-	font-family: 'Roboto Slab', serif;
-}
-.lightBaseTheme {  background: #e9f0f8;  }
-.lightBaseTheme h2{
-	padding: 50px;
-	font-size: 24px;
-	text-align: center;
-	color: #424242;
-}
-.lightBaseTheme p{
-	font-size: 16px;
-	text-align: center;
-	color: #212121;
-	padding: 10px 50px;
-}
-.lightBaseTheme footer{
-	margin-top:100px;
-	padding:20px;
-	color: #212121;
-	background: #f8f9fe;
-}
-.darkBaseTheme {  background: #091720;  }
-.darkBaseTheme h2{
-	padding: 50px;
-	font-size: 24px;
-	text-align: center;
-	color: #b0b0b0;
-}
-.darkBaseTheme p{
-	font-size: 16px;
-	text-align: center;
-	color: #8c9a92;
-	padding: 10px 50px;
-}
-.darkBaseTheme footer{
-	margin-top:100px;
-	padding:20px;
-	color: #8c9a92;
-	background: #162e39;
-}
