+++ samples/public/index.html
@@ -1,12 +1,12 @@
-<!DOCTYPE html>
 <html lang="en">
 
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <title>VizGrammar Charting Samples</title>
+    <title>React-VizGrammar Charting Made Easy</title>
     <link href="https://fonts.googleapis.com/css?family=Roboto+Slab" rel="stylesheet">
+    <!-- <link href="./snippet-hghlight.css" rel="stylesheet"> -->
 </head>
 
 <body>
@@ -14,3 +14,4 @@
     <script src="./app.js"></script>
 </body>
 </html>
+
