+++ samples/styles/snippet-highlight.css
