+++ samples/webpack.config.js
@@ -1,4 +1,4 @@
-const path = require('path');
+path = require('path');
 
 module.exports = {
     devtool: 'source-map',
@@ -39,3 +39,4 @@ module.exports = {
         extensions: ['.js', '.json', '.jsx', '.scss'],
     },
 };
+
